+++ staging/src/k8s.io/component-base/metrics/testutil/BUILD
@@ -2,13 +2,18 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["testutil.go"],
+    srcs = [
+        "metrics.go",
+        "testutil.go",
+    ],
     importmap = "k8s.io/kubernetes/vendor/k8s.io/component-base/metrics/testutil",
     importpath = "k8s.io/component-base/metrics/testutil",
     visibility = ["//visibility:public"],
     deps = [
         "//staging/src/k8s.io/component-base/metrics:go_default_library",
         "//vendor/github.com/prometheus/client_golang/prometheus/testutil:go_default_library",
+        "//vendor/github.com/prometheus/common/expfmt:go_default_library",
+        "//vendor/github.com/prometheus/common/model:go_default_library",
     ],
 )
