+++ test/e2e/framework/metrics/BUILD
@@ -13,8 +13,6 @@ go_library(
         "cluster_autoscaler_metrics.go",
         "controller_manager_metrics.go",
         "e2e_metrics.go",
-        "etcd.go",
-        "generic_metrics.go",
         "interesting_metrics.go",
         "kubelet_metrics.go",
         "latencies.go",
@@ -34,13 +32,12 @@ go_library(
         "//staging/src/k8s.io/apimachinery/pkg/fields:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/sets:go_default_library",
         "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
+        "//staging/src/k8s.io/component-base/metrics/testutil:go_default_library",
         "//test/e2e/framework/log:go_default_library",
         "//test/e2e/framework/ssh:go_default_library",
         "//test/e2e/perftype:go_default_library",
         "//test/e2e/system:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
-        "//vendor/github.com/prometheus/common/expfmt:go_default_library",
-        "//vendor/github.com/prometheus/common/model:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
     ],
 )
