+++ test/e2e/framework/metrics/cluster_autoscaler_metrics.go
@@ -16,22 +16,24 @@ limitations under the License.
 
 package metrics
 
+import "k8s.io/component-base/metrics/testutil"
+
 // ClusterAutoscalerMetrics is metrics for cluster autoscaler
-type ClusterAutoscalerMetrics Metrics
+type ClusterAutoscalerMetrics testutil.Metrics
 
 // Equal returns true if all metrics are the same as the arguments.
 func (m *ClusterAutoscalerMetrics) Equal(o ClusterAutoscalerMetrics) bool {
-	return (*Metrics)(m).Equal(Metrics(o))
+	return (*testutil.Metrics)(m).Equal(testutil.Metrics(o))
 }
 
 func newClusterAutoscalerMetrics() ClusterAutoscalerMetrics {
-	result := NewMetrics()
+	result := testutil.NewMetrics()
 	return ClusterAutoscalerMetrics(result)
 }
 
 func parseClusterAutoscalerMetrics(data string) (ClusterAutoscalerMetrics, error) {
 	result := newClusterAutoscalerMetrics()
-	if err := parseMetrics(data, (*Metrics)(&result)); err != nil {
+	if err := testutil.ParseMetrics(data, (*testutil.Metrics)(&result)); err != nil {
 		return ClusterAutoscalerMetrics{}, err
 	}
 	return result, nil
