+++ test/e2e/framework/metrics/controller_manager_metrics.go
@@ -16,22 +16,24 @@ limitations under the License.
 
 package metrics
 
+import "k8s.io/component-base/metrics/testutil"
+
 // ControllerManagerMetrics is metrics for controller manager
-type ControllerManagerMetrics Metrics
+type ControllerManagerMetrics testutil.Metrics
 
 // Equal returns true if all metrics are the same as the arguments.
 func (m *ControllerManagerMetrics) Equal(o ControllerManagerMetrics) bool {
-	return (*Metrics)(m).Equal(Metrics(o))
+	return (*testutil.Metrics)(m).Equal(testutil.Metrics(o))
 }
 
 func newControllerManagerMetrics() ControllerManagerMetrics {
-	result := NewMetrics()
+	result := testutil.NewMetrics()
 	return ControllerManagerMetrics(result)
 }
 
 func parseControllerManagerMetrics(data string) (ControllerManagerMetrics, error) {
 	result := newControllerManagerMetrics()
-	if err := parseMetrics(data, (*Metrics)(&result)); err != nil {
+	if err := testutil.ParseMetrics(data, (*testutil.Metrics)(&result)); err != nil {
 		return ControllerManagerMetrics{}, err
 	}
 	return result, nil
