--- test/e2e/framework/metrics/generic_metrics.go
@@ -1,81 +1,0 @@
-/*
-Copyright 2015 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package metrics
-
-import (
-	"io"
-	"reflect"
-	"strings"
-
-	"github.com/prometheus/common/expfmt"
-	"github.com/prometheus/common/model"
-	"k8s.io/klog"
-)
-
-// Metrics is generic metrics for other specific metrics
-type Metrics map[string]model.Samples
-
-// Equal returns true if all metrics are the same as the arguments.
-func (m *Metrics) Equal(o Metrics) bool {
-	leftKeySet := []string{}
-	rightKeySet := []string{}
-	for k := range *m {
-		leftKeySet = append(leftKeySet, k)
-	}
-	for k := range o {
-		rightKeySet = append(rightKeySet, k)
-	}
-	if !reflect.DeepEqual(leftKeySet, rightKeySet) {
-		return false
-	}
-	for _, k := range leftKeySet {
-		if !(*m)[k].Equal(o[k]) {
-			return false
-		}
-	}
-	return true
-}
-
-// NewMetrics returns new metrics which are initialized.
-func NewMetrics() Metrics {
-	result := make(Metrics)
-	return result
-}
-
-func parseMetrics(data string, output *Metrics) error {
-	dec := expfmt.NewDecoder(strings.NewReader(data), expfmt.FmtText)
-	decoder := expfmt.SampleDecoder{
-		Dec:  dec,
-		Opts: &expfmt.DecodeOptions{},
-	}
-
-	for {
-		var v model.Vector
-		if err := decoder.Decode(&v); err != nil {
-			if err == io.EOF {
-				// Expected loop termination condition.
-				return nil
-			}
-			klog.Warningf("Invalid Decode. Skipping.")
-			continue
-		}
-		for _, metric := range v {
-			name := string(metric.Metric[model.MetricNameLabel])
-			(*output)[name] = append((*output)[name], metric)
-		}
-	}
-}
